+++ server/src-lib/Hasura/Backends/BigQuery/Types.hs
@@ -34,6 +34,7 @@ module Hasura.Backends.BigQuery.Types
     Reselect (..),
     ScalarType (..),
     Select (..),
+    SelectJson (..),
     TableName (..),
     Time (..),
     Timestamp (..),
@@ -433,6 +434,7 @@ instance NFData a => NFData (Countable a)
 data From
   = FromQualifiedTable (Aliased TableName)
   | FromSelect (Aliased Select)
+  | FromSelectJson (Aliased SelectJson)
   deriving (Eq, Show, Generic, Data, Lift, Ord)
 
 instance FromJSON From
@@ -442,6 +444,20 @@ instance Hashable From
 instance Cacheable From
 
 instance NFData From
+
+data SelectJson = SelectJson
+  { selectJsonBody :: Expression,
+    selectJsonFields :: [(ColumnName, ScalarType)]
+  }
+  deriving (Eq, Ord, Show, Generic, Data, Lift)
+
+instance FromJSON SelectJson
+
+instance Hashable SelectJson
+
+instance Cacheable SelectJson
+
+instance NFData SelectJson
 
 data OpenJson = OpenJson
   { openJsonExpression :: Expression,